refactor: split pipeline into separate modules
